pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fnid-portal"
version = "1.0.0"
description = "FNID Area 3 Operational Portal - Jamaica Constabulary Force"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "Flask>=2.3,<4.0",
    "Flask-Login>=0.6,<1.0",
    "gunicorn>=21.0,<23.0",
    "openpyxl>=3.1,<4.0",
    "python-dotenv>=1.0,<2.0",
    "xhtml2pdf>=0.2.16,<0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py39"
line-length = 120
src = ["fnid_portal", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
